refactor(positions): extract Y.Doc position leaves to src/shared
#1471 will mint CRDT anchors on the client. The mint helper is server-located
but its dependency closure is pure Yjs — except four leaves live inside
`src/server/mcp/document-model.ts`, which imports `node:path` and `saveMarkdown`
at module level. Neither is in the closure, but a client bundle cannot reach
past them, and there is no `@server` alias to borrow. So this is a leaf
extraction establishing a new boundary, not the file move it looks like.

Six functions move to `src/shared/positions/ydoc.ts` with their bodies
untouched. Re-export shims keep every existing import path working: the gate
for this branch was the full suite green with ZERO test-file edits, because a
refactor that forces test changes cannot demonstrate it was behaviour-
preserving. Met — the only test-tree change is two new files.

The three-import ceiling is what keeps `node:path` and the remark pipeline out
of the browser bundle, and it cannot be held by convention: biome's linter is
disabled entirely and there is no no-restricted-imports rule, so a stray import
would land with nothing to object. `ydoc-import-ceiling.test.ts` enforces it,
and walks the transitive closure rather than only direct imports — the direct
ceiling would still pass if `offsets.ts` grew a server import tomorrow.

Adds `anchorFlatRange`, which ships one branch ahead of its first consumer. Its
all-or-nothing contract is not tidiness: the two assocs refuse to mint in
different places around an empty block, so a partially anchored range would have
halves describing different documents.

It is NOT yet the sole assembler of a RelativeRange, and the JSDoc says so
rather than claiming the end state. Collapsing the three existing sites is
deliberately separate: `anchoredRange` re-resolves both endpoints on failure and
logs only when neither was clamped by a heading prefix, so porting that
diagnostic is a precondition rather than a detail.
